Change the type of value for one monitoring point from double to int

+++ include/osquery/numeric_monitoring.h
@@ -25,7 +25,7 @@ namespace monitoring {
 using Clock = std::chrono::system_clock;
 using TimePoint = Clock::time_point;
 
-using ValueType = double;
+using ValueType = long long int;
 
 enum class AggregationType : unsigned {
   None,
